trigger automatic update for k8s cluster
I already did this locally, and it caused an... _incident_.
the original reason was to get access to the metrics config inside the helm chart.
For some reason, I thought this had broken the whole traefik install (which it hadn't) due to failures in mounting chart-values.
The thing is, I never used chart-values before, instead I'm taking the HelmChartConfig route.
Sooo... Everything is good and well, but the upgrade somehow enabled metrics by default.
next up: integrate into prometheus